Use center info as much as possible
